ANN project
Problem Statement: - Create a function approximation model based on a data set. Then Apply various Neural Network architectures based on Radial Basis Functions as RBFN, GRNN, RBFN trained by Bayesian regularization, MLP.
Compare you obtained results with Multilayer perceptron and Linear regression models.
The problem of the international airline passengers prediction problem. This is a problem where given a year and a month, the task is to predict the number of international airline passengers in units of 1000. The data ranges from January 1949 to December 1960 or 12 years, with 144 observations.